chore: inform about semantic versioning
BREAKING CHANGE: Starting with this version, we enforce semantic
versioning. In the future, breaking changes will be pointed out in the
changelog and the major version will be bumped automatically. That also
means that the change of the major version can happen quite frequently
and is not related to major new features.