fix(retrieval/build): improve chunking strategy
The function `cbrkit.retrieval.build` now expects the `chunksize` parameter to be a string literal or an integer.
Previously, `None` was supported, but this has been removed in favor of the explicit literal values.
You can set it to `auto` to derive an optimal size or to `skip` to handle each batch separately
(this can reduce the memory consumption for many cases/queries).